requirements.txt
streamlit==1.38.0
google-generativeai==0.8.1
openai-whisper==20231117
streamlit-mic-recorder==0.0.7